Isolate WASM implementations for tonic_wrap and grpc client
