+++ libs/sdk-common/src/grpc/mod.rs
@@ -1,1 +1,4 @@
 tonic::include_proto!("breez");
+
+#[cfg_attr(target_arch = "wasm32", path = "transport_wasm.rs")]
+pub(crate) mod transport;
